Code cleaning and limit image sizes

+++ app/Console/Commands/SendAttestReminderMail.php
@@ -40,8 +40,8 @@ class SendAttestReminderMail extends Command
     public function handle()
     {
         logger("Running job for sending attest mail reminders!");
-        $previous_month = date("m", strtotime("first day of previous month"));
-        $previous_month_year = date("Y", strtotime("first day of previous month"));
+        $previous_month = (int)date("m", strtotime("first day of previous month"));
+        $previous_month_year = (int)date("Y", strtotime("first day of previous month"));
 
         $amountsent = 0;
         $amountfailed = 0;
