+++ app/Http/helpers.php
@@ -6,8 +6,8 @@ function user_should_attest() {
     $user = Auth::user();
 
     setlocale(LC_TIME, $user->locale_id);
-    $previous_month = date("m", strtotime("first day of previous month"));
-    $previous_month_year = date("Y", strtotime("first day of previous month"));
+    $previous_month = (int)date("m", strtotime("first day of previous month"));
+    $previous_month_year = (int)date("Y", strtotime("first day of previous month"));
 
     $last_month_is_attested = $user->month_is_fully_attested($previous_month_year, $previous_month, 0.5);
